삭제전에 한번 더 물어보도록

+++ www/places/place.js
@@ -30,18 +30,26 @@ angular.module('placekoob.controllers')
   }
 
   place.deletePlace = function() {
-    RemoteAPIService.deleteUserPost(place.uplace_uuid)
-    .then(function() {
-      $ionicPopup.alert({
-        title: '성공',
-        template: '삭제되었습니다'
-      })
-      .then(function() {
-        $ionicHistory.goBack();
-      });
-    }, function(err) {
-      console.error(err);
-    });
+    $ionicPopup.confirm({
+			title: '장소 삭제',
+			template: '정말로 저장한 장소를 지우시겠습니까?'
+		})
+		.then(function(res){
+			if (res) {
+        RemoteAPIService.deleteUserPost(place.uplace_uuid)
+        .then(function() {
+          $ionicPopup.alert({
+            title: '성공',
+            template: '삭제되었습니다'
+          })
+          .then(function() {
+            $ionicHistory.goBack();
+          });
+        }, function(err) {
+          console.error(err);
+        });
+      }
+		});
   }
 
   place.goBack = function() {
@@ -249,7 +257,7 @@ angular.module('placekoob.controllers')
     console.log('Calculated query : ', query);
     query = encodeURI(query);
     console.log('URL encoded query : ', query);
-    
+
     window.open('https://m.search.naver.com/search.naver?sm=mtb_hty.top&where=m_blog&query=' + query, '_system');
   };
